Extract mnemonic input logic for use in HD wallet seed generation
